Guns Implementation
Created two new classes that inherit from a new class called Gun.
AmmoGun is a gun that uses ammo.
Turret is a gun without ammo that has it's own behaviour (not implemented yet)
Now player has two AmmoGun* one for the cannon and one for the laser, and a Turret vector. This components are going to be inactive until player collides with an item that contains one of them.
Now item just creates an instance at the scenary that change it's sprite and it's type(enum) when it recieves a certain amount of damage and when player gets one of them, apply their effects.